+++ lib/alice.dart
@@ -23,9 +23,6 @@ class Alice {
   /// with sensors)
   final bool showInspectorOnShake;
 
-  /// Should inspector use dark theme
-  final bool darkTheme;
-
   /// Icon url for notification
   final String notificationIcon;
 
@@ -50,7 +47,6 @@ class Alice {
     GlobalKey<NavigatorState>? navigatorKey,
     this.showNotification = true,
     this.showInspectorOnShake = false,
-    this.darkTheme = false,
     this.notificationIcon = '@mipmap/ic_launcher',
     this.maxCallsCount = 1000,
     this.directionality,
@@ -61,7 +57,6 @@ class Alice {
       _navigatorKey,
       showNotification: showNotification,
       showInspectorOnShake: showInspectorOnShake,
-      darkTheme: darkTheme,
       notificationIcon: notificationIcon,
       maxCallsCount: maxCallsCount,
       directionality: directionality,