+++ lib/ui/page/alice_call_details_screen.dart
@@ -6,6 +6,7 @@ import 'package:alice/ui/widget/alice_call_overview_widget.dart';
 import 'package:alice/ui/widget/alice_call_request_widget.dart';
 import 'package:alice/ui/widget/alice_call_response_widget.dart';
 import 'package:alice/utils/alice_constants.dart';
+import 'package:alice/utils/alice_theme.dart';
 import 'package:collection/collection.dart' show IterableExtension;
 import 'package:flutter/material.dart';
 import 'package:share_plus/share_plus.dart';
@@ -37,8 +38,7 @@ class _AliceCallDetailsScreenState extends State<AliceCallDetailsScreen>
       textDirection: widget.core.directionality ?? Directionality.of(context),
       child: Theme(
         data: ThemeData(
-          brightness: widget.core.brightness,
-          colorScheme: ColorScheme.light(secondary: AliceConstants.lightRed),
+          colorScheme: AliceTheme.getColorScheme(),
         ),
         child: StreamBuilder<List<AliceHttpCall>>(
           stream: widget.core.callsSubject,