+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.4.0
 * [BREAKING_CHANGE] Updated dart min version to 3.0.0.
+* [BREAKING_CHANGE] Removed `darkTheme` parameter. Alice will now automatically detect the color scheme.
 * Updated dependencies.
 * Fixed lints.
 * Updated example.